third_party/WebKit/LayoutTests/http/tests/security/sandboxed-iframe-meta-refresh.html

<!DOCTYPE html>
<html>
<head>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFrames();
        }
    </script>
</head>

<body>
    <p>Sandboxing a frame puts it into a unique origin by default, which the
    containing document shouldn't have script access to. This test passes if an
    exception is thrown upon the access violation.</p>

    <iframe sandbox
            srcdoc="<head><meta http-equiv='refresh' content='0,http://localhost:8080/security/resources/fail.html'></head><body>PASS</body>"></iframe>
</body>
</html>
